+++ src/main/kotlin/icu/windea/pls/lang/listeners/ParadoxModSettingsListeners.kt
@@ -3,25 +3,25 @@ package icu.windea.pls.lang.listeners
 import com.intellij.openapi.application.*
 import com.intellij.openapi.project.*
 import com.intellij.openapi.roots.*
+import com.intellij.ui.*
 import icu.windea.pls.core.*
 import icu.windea.pls.lang.*
+import icu.windea.pls.lang.editor.*
 import icu.windea.pls.lang.settings.*
 
 /**
- * 当更改游戏配置后，刷新库信息。
+ * 当添加或更改模组配置后，刷新库信息。
  *
  * @see ParadoxLibrary
  * @see ParadoxLibraryProvider
  */
-class ParadoxUpdateLibraryOnGameSettingsChangedListener : ParadoxGameSettingsListener {
-    //目前不考虑onRemove的情况
-
-    override fun onAdd(gameSettings: ParadoxGameSettingsState) {
-        doUpdateLibrary(gameSettings.gameDirectory)
+class ParadoxUpdateLibraryOnModSettingsChangedListener : ParadoxModSettingsListener {
+    override fun onAdd(modSettings: ParadoxModSettingsState) {
+        doUpdateLibrary(modSettings.modDirectory)
     }
 
-    override fun onChange(gameSettings: ParadoxGameSettingsState) {
-        doUpdateLibrary(gameSettings.gameDirectory)
+    override fun onChange(modSettings: ParadoxModSettingsState) {
+        doUpdateLibrary(modSettings.modDirectory)
     }
 
     //org.jetbrains.kotlin.idea.core.script.ucache.ScriptClassRootsUpdater.doUpdate
@@ -35,5 +35,20 @@ class ParadoxUpdateLibraryOnGameSettingsChangedListener : ParadoxGameSettingsLis
             val paradoxLibrary = project.paradoxLibrary
             paradoxLibrary.refreshRoots()
         }
+    }
+}
+
+/**
+ * 当更改模组配置后，更新编辑器通知。
+ *
+ * @see ParadoxGameDirectoryNotConfiguredEditorNotificationProvider
+ */
+class ParadoxUpdateEditorNotificationsOnModSettingsChangedListener : ParadoxModSettingsListener {
+    override fun onAdd(modSettings: ParadoxModSettingsState) {
+
+    }
+
+    override fun onChange(modSettings: ParadoxModSettingsState) {
+        EditorNotifications.updateAll()
     }
 }